Move configuration tests to correct file
